refactor: removing duplicated `ImportMap` struct
