+++ pkg/function/deploy.go
@@ -1,7 +1,6 @@
 package function
 
 import (
-	"bytes"
 	"context"
 	"encoding/json"
 	"fmt"
@@ -9,17 +8,15 @@ import (
 	"io/fs"
 	"mime/multipart"
 	"os"
-	"path"
 	"path/filepath"
-	"regexp"
-	"strings"
 
 	"github.com/go-errors/errors"
+	"github.com/spf13/afero"
+	"github.com/supabase/cli/internal/utils"
 	"github.com/supabase/cli/pkg/api"
 	"github.com/supabase/cli/pkg/cast"
 	"github.com/supabase/cli/pkg/config"
 	"github.com/supabase/cli/pkg/queue"
-	"github.com/tidwall/jsonc"
 )
 
 var ErrNoDeploy = errors.New("All Functions are up to date.")
@@ -166,7 +163,7 @@ func writeForm(form *multipart.Writer, meta api.FunctionDeployMetadata, fsys fs.
 		return nil
 	}
 	// Add import map
-	importMap := ImportMap{}
+	importMap := utils.ImportMap{}
 	if imPath := cast.Val(meta.ImportMapPath, ""); len(imPath) > 0 {
 		data, err := fs.ReadFile(fsys, filepath.FromSlash(imPath))
 		if err != nil {
@@ -175,9 +172,13 @@ func writeForm(form *multipart.Writer, meta api.FunctionDeployMetadata, fsys fs.
 		if err := importMap.Parse(data); err != nil {
 			return err
 		}
-		if err := importMap.Resolve(imPath, fsys); err != nil {
+
+		// Casting io.FS to afero.Fs
+		afsys := &afero.FromIOFS{FS: fsys}
+		if err := importMap.Resolve(imPath, afsys); err != nil {
 			return err
 		}
+
 		// TODO: replace with addFile once edge runtime supports jsonc
 		fmt.Fprintf(os.Stderr, "Uploading asset (%s): %s\n", *meta.Name, imPath)
 		f, err := form.CreateFormFile("file", imPath)
@@ -199,123 +200,5 @@ func writeForm(form *multipart.Writer, meta api.FunctionDeployMetadata, fsys fs.
 			return err
 		}
 	}
-	return walkImportPaths(meta.EntrypointPath, importMap, addFile)
-}
-
-type ImportMap struct {
-	Imports map[string]string            `json:"imports"`
-	Scopes  map[string]map[string]string `json:"scopes"`
-}
-
-func (m *ImportMap) Parse(data []byte) error {
-	data = jsonc.ToJSONInPlace(data)
-	decoder := json.NewDecoder(bytes.NewReader(data))
-	if err := decoder.Decode(&m); err != nil {
-		return errors.Errorf("failed to parse import map: %w", err)
-	}
-	return nil
-}
-
-func (m *ImportMap) Resolve(imPath string, fsys fs.FS) error {
-	// Resolve all paths relative to current file
-	for k, v := range m.Imports {
-		m.Imports[k] = resolveHostPath(imPath, v, fsys)
-	}
-	for module, mapping := range m.Scopes {
-		for k, v := range mapping {
-			m.Scopes[module][k] = resolveHostPath(imPath, v, fsys)
-		}
-	}
-	return nil
-}
-
-func resolveHostPath(jsonPath, hostPath string, fsys fs.FS) string {
-	// Leave absolute paths unchanged
-	if path.IsAbs(hostPath) {
-		return hostPath
-	}
-	resolved := path.Join(path.Dir(jsonPath), hostPath)
-	if _, err := fs.Stat(fsys, filepath.FromSlash(resolved)); err != nil {
-		// Leave URLs unchanged
-		return hostPath
-	}
-	// Directory imports need to be suffixed with /
-	// Ref: https://deno.com/manual@v1.33.0/basics/import_maps
-	if strings.HasSuffix(hostPath, "/") {
-		resolved += "/"
-	}
-	// Relative imports must be prefixed with ./ or ../
-	if !path.IsAbs(resolved) {
-		resolved = "./" + resolved
-	}
-	return resolved
-}
-
-// Ref: https://regex101.com/r/DfBdJA/1
-var importPathPattern = regexp.MustCompile(`(?i)(?:import|export)\s+(?:{[^{}]+}|.*?)\s*(?:from)?\s*['"](.*?)['"]|import\(\s*['"](.*?)['"]\)`)
-
-func walkImportPaths(srcPath string, importMap ImportMap, readFile func(curr string, w io.Writer) error) error {
-	seen := map[string]struct{}{}
-	// DFS because it's more efficient to pop from end of array
-	q := make([]string, 1)
-	q[0] = srcPath
-	for len(q) > 0 {
-		curr := q[len(q)-1]
-		q = q[:len(q)-1]
-		// Assume no file is symlinked
-		if _, ok := seen[curr]; ok {
-			continue
-		}
-		seen[curr] = struct{}{}
-		// Read into memory for regex match later
-		var buf bytes.Buffer
-		if err := readFile(curr, &buf); errors.Is(err, os.ErrNotExist) {
-			fmt.Fprintln(os.Stderr, "WARN:", err)
-			continue
-		} else if err != nil {
-			return err
-		}
-		// Traverse all modules imported by the current source file
-		for _, matches := range importPathPattern.FindAllStringSubmatch(buf.String(), -1) {
-			if len(matches) < 3 {
-				continue
-			}
-			// Matches 'from' clause if present, else fallback to 'import'
-			mod := matches[1]
-			if len(mod) == 0 {
-				mod = matches[2]
-			}
-			mod = strings.TrimSpace(mod)
-			// Substitute kv from import map
-			substituted := false
-			for k, v := range importMap.Imports {
-				if strings.HasPrefix(mod, k) {
-					mod = v + mod[len(k):]
-					substituted = true
-				}
-			}
-			// Ignore URLs and directories
-			if len(path.Ext(mod)) == 0 {
-				continue
-			}
-			// Deno import path must begin with one of these prefixes
-			if !isRelPath(mod) && !isAbsPath(mod) {
-				continue
-			}
-			if isRelPath(mod) && !substituted {
-				mod = path.Join(path.Dir(curr), mod)
-			}
-			// Cleans import path to help detect duplicates
-			q = append(q, path.Clean(mod))
-		}
-	}
-	return nil
-}
-
-func isRelPath(mod string) bool {
-	return strings.HasPrefix(mod, "./") || strings.HasPrefix(mod, "../")
-}
-
-func isAbsPath(mod string) bool {
-	return strings.HasPrefix(mod, "/")
+	return importMap.WalkImportPaths(meta.EntrypointPath, addFile)
 }
